requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
django-ratelimit>=4.0,<5.0